@@ -1,10 +1,10 @@
 name: flutter_webrtc
 description: Flutter WebRTC plugin for iOS/Android/Destkop/Web, based on GoogleWebRTC.
-version: 0.5.8
+version: 0.6.0-nullsafety.0
 homepage: https://github.com/cloudwebrtc/flutter-webrtc
 environment:
-  sdk: '>=2.2.2 <3.0.0'
-  flutter: ^1.10.0
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ^1.22.0
 
 dependencies:
   flutter:
