@@ -51,7 +51,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  List<RouteItem> items;
+  late List<RouteItem> items;
 
   @override
   void initState() {
