@@ -121,9 +121,9 @@ abstract class MediaDevices {
 class MediaDeviceInfo {
   MediaDeviceInfo({
     this.kind,
-    this.label,
+    required this.label,
     this.groupId,
-    this.deviceId,
+    required this.deviceId,
   });
 
   /// Returns a String that is an identifier for the represented device that
@@ -136,11 +136,11 @@ class MediaDeviceInfo {
   /// Returns a String that is a group identifier. Two devices have the same
   /// group identifier if they belong to the same physical device
   /// — for example a monitor with both a built-in camera and a microphone.
-  final String groupId;
+  final String? groupId;
 
   /// Returns an enumerated value that is either 'videoinput', 'audioinput' or
   /// 'audiooutput'.
-  final String kind;
+  final String? kind;
 
   /// Returns a String that is a label describing this device
   /// (for example "External USB Webcam").
