@@ -19,22 +19,22 @@ class RTCRTPCodec {
         parameters: map['parameters']);
   }
   // Payload type used to identify this codec in RTP packets.
-  int payloadType;
+  int? payloadType;
 
   /// Name used to identify the codec. Equivalent to MIME subtype.
-  String name;
+  String? name;
 
   /// The media type of this codec. Equivalent to MIME top-level type.
-  String kind;
+  String? kind;
 
   /// Clock rate in Hertz.
-  int clockRate;
+  int? clockRate;
 
   /// The number of audio channels used. Set to null for video codecs.
-  int numChannels;
+  int? numChannels;
 
   /// The "format specific parameters" field from the "a=fmtp" line in the SDP
-  Map<dynamic, dynamic> parameters;
+  Map<dynamic, dynamic>? parameters;
 
   Map<String, dynamic> toMap() {
     return {
@@ -49,15 +49,16 @@ class RTCRTPCodec {
 }
 
 class RTCRtpEncoding {
-  RTCRtpEncoding(
-      {this.rid,
-      this.active,
-      this.maxBitrate,
-      this.maxFramerate,
-      this.minBitrate,
-      this.numTemporalLayers,
-      this.scaleResolutionDownBy,
-      this.ssrc});
+  RTCRtpEncoding({
+    this.rid,
+    this.active = true,
+    this.maxBitrate,
+    this.maxFramerate,
+    this.minBitrate,
+    this.numTemporalLayers = 1,
+    this.scaleResolutionDownBy = 1.0,
+    this.ssrc,
+  });
 
   factory RTCRtpEncoding.fromMap(Map<dynamic, dynamic> map) {
     return RTCRtpEncoding(
@@ -73,44 +74,43 @@ class RTCRtpEncoding {
 
   /// If non-null, this represents the RID that identifies this encoding layer.
   /// RIDs are used to identify layers in simulcast.
-  String rid;
+  String? rid;
 
   /// Set to true to cause this encoding to be sent, and false for it not to
   /// be sent.
-  bool active = true;
+  bool active;
 
   /// If non-null, this represents the Transport Independent Application
   /// Specific maximum bandwidth defined in RFC3890. If null, there is no
   /// maximum bitrate.
-  int maxBitrate;
+  int? maxBitrate;
 
   /// The minimum bitrate in bps for video.
-  int minBitrate;
+  int? minBitrate;
 
   /// The max framerate in fps for video.
-  int maxFramerate;
+  int? maxFramerate;
 
   /// The number of temporal layers for video.
-  int numTemporalLayers = 1;
+  int? numTemporalLayers;
 
   /// If non-null, scale the width and height down by this factor for video. If null,
   /// implementation default scaling factor will be used.
-  double scaleResolutionDownBy = 1.0;
+  double? scaleResolutionDownBy;
 
   /// SSRC to be used by this encoding.
   /// Can't be changed between getParameters/setParameters.
-  int ssrc;
+  int? ssrc;
 
   Map<String, dynamic> toMap() {
     return {
       if (rid != null) 'rid': rid,
-      'active': active ?? true,
+      'active': active,
       if (maxBitrate != null) 'maxBitrate': maxBitrate,
       if (maxFramerate != null) 'maxFramerate': maxFramerate,
       if (minBitrate != null) 'minBitrate': minBitrate,
-      if (numTemporalLayers != null) 'numTemporalLayers': numTemporalLayers,
-      if (scaleResolutionDownBy != null)
-        'scaleResolutionDownBy': scaleResolutionDownBy,
+      'numTemporalLayers': numTemporalLayers,
+      'scaleResolutionDownBy': scaleResolutionDownBy,
       if (ssrc != null) 'ssrc': ssrc,
     };
   }
@@ -124,13 +124,13 @@ class RTCHeaderExtension {
   }
 
   /// The URI of the RTP header extension, as defined in RFC5285.
-  String uri;
+  String? uri;
 
   /// The value put in the RTP packet to identify the header extension.
-  int id;
+  int? id;
 
   /// Whether the header extension is encrypted or not.
-  bool encrypted;
+  bool? encrypted;
 
   Map<String, dynamic> toMap() {
     return {
@@ -142,12 +142,13 @@ class RTCHeaderExtension {
 }
 
 class RTCRtpParameters {
-  RTCRtpParameters(
-      {this.transactionId,
-      this.rtcp,
-      this.headerExtensions,
-      this.encodings,
-      this.codecs});
+  RTCRtpParameters({
+    this.transactionId,
+    this.rtcp,
+    this.headerExtensions,
+    this.encodings,
+    this.codecs,
+  });
 
   factory RTCRtpParameters.fromMap(Map<dynamic, dynamic> map) {
     var encodings = <RTCRtpEncoding>[];
@@ -174,35 +175,35 @@ class RTCRtpParameters {
         codecs: codecs);
   }
 
-  String transactionId;
+  String? transactionId;
 
-  RTCRTCPParameters rtcp;
+  RTCRTCPParameters? rtcp;
 
-  List<RTCHeaderExtension> headerExtensions;
+  List<RTCHeaderExtension>? headerExtensions;
 
-  List<RTCRtpEncoding> encodings;
+  List<RTCRtpEncoding>? encodings;
 
   /// Codec parameters can't currently be changed between getParameters and
   /// setParameters. Though in the future it will be possible to reorder them or
   /// remove them.
-  List<RTCRTPCodec> codecs;
+  List<RTCRTPCodec>? codecs;
 
   Map<String, dynamic> toMap() {
     var headerExtensionsList = <dynamic>[];
-    headerExtensions.forEach((params) {
+    headerExtensions?.forEach((params) {
       headerExtensionsList.add(params.toMap());
     });
     var encodingList = <dynamic>[];
-    encodings.forEach((params) {
+    encodings?.forEach((params) {
       encodingList.add(params.toMap());
     });
     var codecsList = <dynamic>[];
-    codecs.forEach((params) {
+    codecs?.forEach((params) {
       codecsList.add(params.toMap());
     });
     return {
       'transactionId': transactionId,
-      'rtcp': rtcp.toMap(),
+      if (rtcp != null) 'rtcp': rtcp!.toMap(),
       'headerExtensions': headerExtensionsList,
       'encodings': encodingList,
       'codecs': codecsList,
