@@ -1,7 +1,7 @@
 class RTCSessionDescription {
   RTCSessionDescription(this.sdp, this.type);
-  String sdp;
-  String type;
+  String? sdp;
+  String? type;
   dynamic toMap() {
     return {'sdp': sdp, 'type': type};
   }
