@@ -4,9 +4,14 @@ import 'rtc_rtp_receiver.dart';
 import 'rtc_rtp_transceiver.dart';
 
 class RTCTrackEvent {
-  RTCTrackEvent({this.receiver, this.streams, this.track, this.transceiver});
-  final RTCRtpReceiver receiver;
+  RTCTrackEvent({
+    this.receiver,
+    required this.streams,
+    required this.track,
+    this.transceiver,
+  });
+  final RTCRtpReceiver? receiver;
   final List<MediaStream> streams;
   final MediaStreamTrack track;
-  final RTCRtpTransceiver transceiver;
+  final RTCRtpTransceiver? transceiver;
 }
