@@ -37,15 +37,15 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   // private:
   final String _peerConnectionId;
   final _channel = WebRTC.methodChannel();
-  StreamSubscription<dynamic> _eventSubscription;
+  StreamSubscription<dynamic>? _eventSubscription;
   final _localStreams = <MediaStream>[];
   final _remoteStreams = <MediaStream>[];
-  RTCDataChannelNative _dataChannel;
+  RTCDataChannelNative? _dataChannel;
   Map<String, dynamic> _configuration;
-  RTCSignalingState _signalingState;
-  RTCIceGatheringState _iceGatheringState;
-  RTCIceConnectionState _iceConnectionState;
-  RTCPeerConnectionState _connectionState;
+  RTCSignalingState? _signalingState;
+  RTCIceGatheringState? _iceGatheringState;
+  RTCIceConnectionState? _iceConnectionState;
+  RTCPeerConnectionState? _connectionState;
 
   final Map<String, dynamic> defaultSdpConstraints = {
     'mandatory': {
@@ -56,16 +56,16 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   };
 
   @override
-  RTCSignalingState get signalingState => _signalingState;
+  RTCSignalingState? get signalingState => _signalingState;
 
   @override
-  RTCIceGatheringState get iceGatheringState => _iceGatheringState;
+  RTCIceGatheringState? get iceGatheringState => _iceGatheringState;
 
   @override
-  RTCIceConnectionState get iceConnectionState => _iceConnectionState;
+  RTCIceConnectionState? get iceConnectionState => _iceConnectionState;
 
   @override
-  RTCPeerConnectionState get connectionState => _connectionState;
+  RTCPeerConnectionState? get connectionState => _connectionState;
 
   Future<RTCSessionDescription> get localDescription => getLocalDescription();
 
@@ -80,19 +80,19 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
     switch (map['event']) {
       case 'signalingState':
         _signalingState = signalingStateForString(map['state']);
-        onSignalingState?.call(_signalingState);
+        onSignalingState?.call(_signalingState!);
         break;
       case 'peerConnectionState':
         _connectionState = peerConnectionStateForString(map['state']);
-        onConnectionState?.call(_connectionState);
+        onConnectionState?.call(_connectionState!);
         break;
       case 'iceGatheringState':
         _iceGatheringState = iceGatheringStateforString(map['state']);
-        onIceGatheringState?.call(_iceGatheringState);
+        onIceGatheringState?.call(_iceGatheringState!);
         break;
       case 'iceConnectionState':
         _iceConnectionState = iceConnectionStateForString(map['state']);
-        onIceConnectionState?.call(_iceConnectionState);
+        onIceConnectionState?.call(_iceConnectionState!);
         break;
       case 'onCandidate':
         Map<dynamic, dynamic> cand = map['candidate'];
@@ -115,10 +115,7 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
         break;
       case 'onRemoveStream':
         String streamId = map['streamId'];
-        var stream =
-            _remoteStreams.firstWhere((it) => it.id == streamId, orElse: () {
-          return null;
-        });
+        var stream = _remoteStreams.firstWhere((it) => it.id == streamId);
         onRemoveStream?.call(stream);
         _remoteStreams.removeWhere((it) => it.id == streamId);
         break;
@@ -151,10 +148,7 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
         break;
       case 'onRemoveTrack':
         String streamId = map['streamId'];
-        var stream =
-            _remoteStreams.firstWhere((it) => it.id == streamId, orElse: () {
-          return null;
-        });
+        var stream = _remoteStreams.firstWhere((it) => it.id == streamId);
         Map<dynamic, dynamic> track = map['track'];
         var oldTrack = MediaStreamTrackNative(
             map['trackId'], track['label'], track['kind'], track['enabled']);
@@ -165,7 +159,7 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
         String label = map['label'];
         _dataChannel =
             RTCDataChannelNative(_peerConnectionId, label, dataChannelId);
-        onDataChannel?.call(_dataChannel);
+        onDataChannel?.call(_dataChannel!);
         break;
       case 'onRenegotiationNeeded':
         onRenegotiationNeeded?.call();
@@ -216,8 +210,7 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   }
 
   void errorListener(Object obj) {
-    final PlatformException e = obj;
-    throw e;
+    if (obj is Exception) throw obj;
   }
 
   @override
@@ -251,10 +244,10 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
 
   @override
   Future<RTCSessionDescription> createOffer(
-      [Map<String, dynamic> constraints = const {}]) async {
+      [Map<String, dynamic>? constraints]) async {
     try {
-      final response = await _channel
-          .invokeMethod<Map<dynamic, dynamic>>('createOffer', <String, dynamic>{
+      final response =
+          await WebRTC.invokeMethod('createOffer', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'constraints': constraints ?? defaultSdpConstraints
       });
@@ -269,13 +262,14 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
 
   @override
   Future<RTCSessionDescription> createAnswer(
-      [Map<String, dynamic> constraints]) async {
+      [Map<String, dynamic>? constraints]) async {
     try {
-      final response = await _channel.invokeMethod<Map<dynamic, dynamic>>(
-          'createAnswer', <String, dynamic>{
+      final response =
+          await WebRTC.invokeMethod('createAnswer', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'constraints': constraints ?? defaultSdpConstraints
       });
+
       String sdp = response['sdp'];
       String type = response['type'];
       return RTCSessionDescription(sdp, type);
@@ -329,10 +323,11 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   @override
   Future<RTCSessionDescription> getLocalDescription() async {
     try {
-      final response = await _channel.invokeMethod<Map<dynamic, dynamic>>(
-          'getLocalDescription', <String, dynamic>{
+      final response =
+          await WebRTC.invokeMethod('getLocalDescription', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
       });
+
       String sdp = response['sdp'];
       String type = response['type'];
       return RTCSessionDescription(sdp, type);
@@ -344,10 +339,11 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   @override
   Future<RTCSessionDescription> getRemoteDescription() async {
     try {
-      final response = await _channel.invokeMethod<Map<dynamic, dynamic>>(
-          'getRemoteDescription', <String, dynamic>{
+      final response =
+          await WebRTC.invokeMethod('getRemoteDescription', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
       });
+
       String sdp = response['sdp'];
       String type = response['type'];
       return RTCSessionDescription(sdp, type);
@@ -365,13 +361,13 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   }
 
   @override
-  Future<List<StatsReport>> getStats([MediaStreamTrack track]) async {
+  Future<List<StatsReport>> getStats([MediaStreamTrack? track]) async {
     try {
-      final response = await _channel.invokeMethod<Map<dynamic, dynamic>>(
-          'getStats', <String, dynamic>{
+      final response = await WebRTC.invokeMethod('getStats', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'track': track != null ? track.id : null
       });
+
       var stats = <StatsReport>[];
       if (response != null) {
         List<dynamic> reports = response['stats'];
@@ -408,7 +404,7 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
       });
       _dataChannel =
           RTCDataChannelNative(_peerConnectionId, label, dataChannelDict.id);
-      return _dataChannel;
+      return _dataChannel!;
     } on PlatformException catch (e) {
       throw 'Unable to RTCPeerConnection::createDataChannel: ${e.message}';
     }
@@ -469,13 +465,13 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
 
   @override
   Future<RTCRtpSender> addTrack(MediaStreamTrack track,
-      [MediaStream stream]) async {
+      [MediaStream? stream]) async {
     try {
       final response =
           await _channel.invokeMethod('addTrack', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'trackId': track.id,
-        'streamIds': [stream.id]
+        'streamIds': [stream?.id]
       });
       return RTCRtpSenderNative.fromMap(response,
           peerConnectionId: _peerConnectionId);
@@ -501,9 +497,9 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
 
   @override
   Future<RTCRtpTransceiver> addTransceiver(
-      {MediaStreamTrack track,
-      RTCRtpMediaType kind,
-      RTCRtpTransceiverInit init}) async {
+      {MediaStreamTrack? track,
+      RTCRtpMediaType? kind,
+      RTCRtpTransceiverInit? init}) async {
     try {
       final response =
           await _channel.invokeMethod('addTransceiver', <String, dynamic>{
