@@ -13,28 +13,24 @@ class MediaRecorderNative extends MediaRecorder {
 
   @override
   Future<void> start(String path,
-      {MediaStreamTrack videoTrack, RecorderAudioChannel audioChannel
+      {MediaStreamTrack? videoTrack, RecorderAudioChannel? audioChannel
       // TODO(cloudwebrtc): add codec/quality options
       }) async {
-    if (path == null) {
-      throw ArgumentError.notNull('path');
-    }
-
     if (audioChannel == null && videoTrack == null) {
       throw Exception('Neither audio nor video track were provided');
     }
 
     await WebRTC.methodChannel().invokeMethod('startRecordToFile', {
       'path': path,
-      'audioChannel': audioChannel?.index,
-      'videoTrackId': videoTrack?.id,
+      if (audioChannel != null) 'audioChannel': audioChannel.index,
+      if (videoTrack != null) 'videoTrackId': videoTrack.id,
       'recorderId': _recorderId
     });
   }
 
   @override
   void startWeb(MediaStream stream,
-      {Function(dynamic blob, bool isLastOne) onDataChunk, String mimeType}) {
+      {Function(dynamic blob, bool isLastOne)? onDataChunk, String? mimeType}) {
     throw 'It\'s for Flutter Web only';
   }
 
