@@ -8,12 +8,12 @@ void main() {
   final channel = WebRTC.methodChannel();
   setUp(() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      await ServicesBinding.instance.defaultBinaryMessenger
+      await ServicesBinding.instance?.defaultBinaryMessenger
           .handlePlatformMessage(
-              'FlutterWebRTC/peerConnectoinEvent', null, (ByteData data) {});
-      await ServicesBinding.instance.defaultBinaryMessenger
+              'FlutterWebRTC/peerConnectoinEvent', null, (ByteData? data) {});
+      await ServicesBinding.instance?.defaultBinaryMessenger
           .handlePlatformMessage(
-              'FlutterWebRTC/dataChannelEvent', null, (ByteData data) {});
+              'FlutterWebRTC/dataChannelEvent', null, (ByteData? data) {});
     });
   });
 
